@@ -5,6 +5,7 @@ import jsPDF from 'jspdf';
 import html2canvas from 'html2canvas';
 import { getLocaleDateFormat } from '@angular/common';
 import { Data } from '@angular/router';
+import * as FileSaver from 'file-saver';
 
 @Component({
   selector: 'app-dados',
@@ -17,16 +18,14 @@ export class DadosComponent implements OnInit {
   constructor(private usuarioService:UsuarioService) { }
 
   ngOnInit(): void {
-    if(sessionStorage.getItem("token")){
-      this.listarUsuarios();
-    }
+    this.listarUsuarios();
   }
 
   listarUsuarios(){
     this.usuarioService.listarUsuarios().subscribe(data => this.usuarios = data);
   }
 
-  openPDF(): void {
+  exportPDF(): void {
     const dataAutal = new Date().toLocaleString();
     let DATA: any = document.getElementById('htmlData');
     html2canvas(DATA).then((canvas) => {
@@ -41,4 +40,6 @@ export class DadosComponent implements OnInit {
       PDF.save('lista-usuario.pdf');
     });
   }
+
 }
+
