@@ -9,16 +9,14 @@ import org.springframework.security.core.userdetails.UserDetails;
 
 import com.login.model.Usuario;
 
-public class DetalheUsuarioData implements UserDetails {
+public class DetalheUsuarioData implements UserDetails{
 	
-private static final long serialVersionUID = 1L;
-	
-	private final Optional<Usuario> usuario;
+	private Optional<Usuario> usuario;
 	
 	public DetalheUsuarioData(Optional<Usuario> usuario) {
 		this.usuario = usuario;
 	}
-	
+
 	@Override
 	public Collection<? extends GrantedAuthority> getAuthorities() {
 		return new ArrayList<>();
@@ -36,25 +34,21 @@ private static final long serialVersionUID = 1L;
 
 	@Override
 	public boolean isAccountNonExpired() {
-		// TODO Auto-generated method stub
 		return true;
 	}
 
 	@Override
 	public boolean isAccountNonLocked() {
-		// TODO Auto-generated method stub
 		return true;
 	}
 
 	@Override
 	public boolean isCredentialsNonExpired() {
-		// TODO Auto-generated method stub
 		return true;
 	}
 
 	@Override
 	public boolean isEnabled() {
-		// TODO Auto-generated method stub
 		return true;
 	}
 
