@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
 public class NotFoundException extends RuntimeException {
-
+	
 	private static final long serialVersionUID = 1L;
 	
 	public NotFoundException () {
@@ -14,6 +14,5 @@ public class NotFoundException extends RuntimeException {
 	public NotFoundException (String texto) {
 		super(texto);
 	}
-
 
 }
