@@ -24,11 +24,11 @@ import com.login.model.Usuario;
 public class JWTAutenticarFilter extends UsernamePasswordAuthenticationFilter {
 	
 	public static final int TOKEN_EXPIRACAO = 900_000;
-	public static final String TOKEN_SENHA = "026cb4eb-08cd-4e9e-90fb-ac23fbbb7c8a";
+	public static final String TOKEN_SENHA = "6609ebd3-b3a8-45d7-a5b6-15ec122c8b25";
 	
 	private final AuthenticationManager authenticationManager;
-	
-	public JWTAutenticarFilter (AuthenticationManager authenticationManager) {
+
+	public JWTAutenticarFilter(AuthenticationManager authenticationManager) {
 		this.authenticationManager = authenticationManager;
 	}
 	
@@ -37,32 +37,31 @@ public class JWTAutenticarFilter extends UsernamePasswordAuthenticationFilter {
 			throws AuthenticationException {
 		
 		try {
-		
 			Usuario usuario = new ObjectMapper().readValue(request.getInputStream(), Usuario.class);
-			
 			return authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(
-					usuario.getEmail(), usuario.getSenha(), new ArrayList<>()
-				));
+					usuario.getEmail(),
+					usuario.getSenha(),
+					new ArrayList<>()
+					));
 		} catch (IOException e) {
-			throw new RuntimeException("Falha ao autenticar usuario", e);
-			
-		}
+			throw new RuntimeException("Falha ao autenticar Usuario");
+		}			
 	}
 	
 	@Override
-	protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain,
-		Authentication authResult) throws IOException, ServletException {
+    protected void successfulAuthentication(HttpServletRequest request,
+                                            HttpServletResponse response,
+                                            FilterChain chain,
+                                            Authentication authResult) throws IOException, ServletException {
+		
 		DetalheUsuarioData usuarioData = (DetalheUsuarioData) authResult.getPrincipal();
 		
-		String token = JWT.create()
-				.withSubject(usuarioData.getUsername())
+		String token = JWT.create().withSubject(usuarioData.getUsername())
 				.withExpiresAt(new Date(System.currentTimeMillis() + TOKEN_EXPIRACAO))
 				.sign(Algorithm.HMAC512(TOKEN_SENHA));
 		
 		response.addHeader("Access-Control-Allow-Origin", "*");
 		response.getWriter().write(token);
 		response.getWriter().flush();
-		
 	}
-
 }
